Fix AutoOrientProcessor
